dashboard/app: show manager unique coverage
1. Make heatmap testable, move out the spanner client instantiation.
2. Generate spannerdb.ReadOnlyTransaction mocks.
3. Generate spannerdb.RowIterator mocks.
4. Generate spannerdb.Row mocks.
5. Prepare spannerdb fixture.
6. Fixed html control name + value.
7. Added multiple tests.
8. Show line coverage from selected manager.
9. Propagate coverage url params to file coverage url.

+++ pkg/covermerger/file_line_merger.go
@@ -20,7 +20,7 @@ func makeFileLineCoverMerger(fvs fileVersions, base RepoCommit) FileCoverageMerg
 	}
 	a := &FileLineCoverMerger{
 		MergeResult: &MergeResult{
-			HitCounts:   make(map[int]int),
+			HitCounts:   make(map[int]int64),
 			FileExists:  true,
 			LineDetails: make(map[int][]*FileRecord),
 		},
@@ -49,7 +49,7 @@ func (a *FileLineCoverMerger) Add(record *FileRecord) {
 		return
 	}
 	if targetLine := a.matchers[record.RepoCommit].SameLinePos(record.StartLine); targetLine != -1 {
-		a.HitCounts[targetLine] += record.HitCount
+		a.HitCounts[targetLine] += int64(record.HitCount)
 		a.LineDetails[targetLine] = append(a.LineDetails[targetLine], record)
 	}
 }
